requirements.txt
absl-py==2.2.2
alembic==1.16.4
altair==5.5.0
amqp==5.3.1
aniso8601==10.0.1
astunparse==1.6.3
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
capstone==5.0.6
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cryptography==43.0.3
cycler==0.12.1
debtcollector==3.0.0
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
dogpile.cache==1.4.0
elementpath==4.8.0
eventlet==0.40.2
fasteners==0.20
flake8==7.3.0
Flask==3.1.2
Flask-RESTful==0.3.10
flatbuffers==25.2.10
fonttools==4.57.0
fusepy==3.0.1
futurist==3.2.0
gast==0.6.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.2
keystone==27.0.0
keystone-engine==0.9.2
keystoneauth1==5.12.0
keystonemiddleware==10.12.0
kiwisolver==1.4.8
kombu==5.5.4
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.1
mock==5.2.0
msgpack==1.1.1
namex==0.0.8
narwhals==1.35.0
netaddr==1.3.0
networkx==3.5
numpy==2.1.3
oauthlib==3.3.1
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.15.0
os-service-types==1.8.0
oslo.cache==3.11.0
oslo.concurrency==7.1.0
oslo.config==10.0.0
oslo.context==6.0.0
oslo.db==17.4.0
oslo.i18n==6.5.1
oslo.log==7.2.0
oslo.messaging==17.1.0
oslo.metrics==0.12.0
oslo.middleware==6.6.0
oslo.policy==4.5.1
oslo.serialization==5.7.0
oslo.service==4.2.2
oslo.upgradecheck==2.5.0
oslo.utils==9.0.0
osprofiler==4.3.0
packaging==24.2
pandas==2.2.3
Paste==3.10.1
PasteDeploy==3.1.0
pbr==7.0.1
pillow==11.2.1
prettytable==3.16.0
prometheus_client==0.22.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pycadf==4.0.1
pycodestyle==2.14.0
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==24.2.1
pyparsing==3.2.3
pysaml2==7.5.2
pyserial==3.5
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dateutil==2.9.0.post0
python-keystoneclient==5.6.0
pytz==2025.2
pyusb==1.3.1
PyYAML==6.0.2
referencing==0.36.2
repoze.lru==0.7
requests==2.32.3
rfc3986==2.0.0
rich==14.0.0
Routes==2.5.1
rpds-py==0.24.0
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
scrypt==0.9.4
shiboken6==6.9.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.43
statsd==4.0.1
stevedore==5.4.1
streamlit==1.44.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
testresources==2.0.2
testscenarios==0.5.0
testtools==2.7.2
threadpoolctl==3.6.0
tifffile==2025.6.11
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
unicorn==2.1.3
urllib3==2.4.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
WebOb==1.8.9
Werkzeug==3.1.3
wrapt==1.17.2
xmlschema==2.5.1
yappi==1.6.10
